ArticaleCode/requirements.txt
pandas==1.2.3
numpy==1.20.2
scikit-learn==0.24.1
matplotlib==3.4.0
seaborn==0.11.1
opencv-python==4.5.3.56
TPOT==0.11.7
slots==0.4.0
auto-sklearn
scipy